+++ gaios/Services/ResolverManager.swift
@@ -12,7 +12,7 @@ class ResolverManager {
     
     init(
         _ factor: TwoFactorCall?,
-        chain: String,
+        network: NetworkSecurityCase,
         connected: @escaping() -> Bool = { true },
         hwDevice: HWProtocol?,
         session: SessionManager? = nil,
@@ -20,11 +20,12 @@ class ResolverManager {
         self.session = session
         resolver = GDKResolver(
             factor,
+            gdkSession: session?.session,
             popupDelegate: PopupResolver(),
             hwDelegate: HWResolver(),
             bcurDelegate: bcurResolver,
             hwDevice: hwDevice,
-            chain: chain,
+            network: network,
             connected: connected
         )
     }
@@ -60,17 +61,17 @@ class PopupResolver: NSObject, UITextFieldDelegate, PopupResolverDelegate {
     
     private var textContinuation: CheckedContinuation<String, Error>?
     
-    func code(_ method: String, attemptsRemaining: Int?) async throws -> String {
+    func code(_ method: String, attemptsRemaining: Int?, enable2faCallMethod: Bool, network: NetworkSecurityCase) async throws -> String {
         DispatchQueue.main.async {
             UIApplication.topViewController()?.stopAnimating()
         }
         return try await withCheckedThrowingContinuation { continuation in
             textContinuation = continuation
-            codeCustomDialog(method, attemptsRemaining: attemptsRemaining)
+            codeCustomDialog(method, attemptsRemaining: attemptsRemaining, enable2faCallMethod: enable2faCallMethod, network: network)
         }
     }
 
-    func codeCustomDialog(_ method: String, attemptsRemaining: Int?) {
+    func codeCustomDialog(_ method: String, attemptsRemaining: Int?, enable2faCallMethod: Bool, network: NetworkSecurityCase) {
         let methodDesc: String
         if method == TwoFactorType.email.rawValue { methodDesc = "id_email" } else if method == TwoFactorType.phone.rawValue { methodDesc = "id_phone_call" } else if method == TwoFactorType.sms.rawValue { methodDesc = "id_sms" } else { methodDesc = "id_authenticator_app" }
         
@@ -79,6 +80,7 @@ class PopupResolver: NSObject, UITextFieldDelegate, PopupResolverDelegate {
             
         vc.commontitle = String(format: NSLocalizedString("id_please_provide_your_1s_code", comment: ""), NSLocalizedString(methodDesc, comment: ""))
         vc.attemptsRemaining = attemptsRemaining ?? 3
+        vc.enable2faCallMethod = enable2faCallMethod
         
         vc.onCancel = { [weak self] in
             if let appDelegate = UIApplication.shared.delegate as? AppDelegate{
@@ -96,7 +98,22 @@ class PopupResolver: NSObject, UITextFieldDelegate, PopupResolverDelegate {
                 UIApplication.topViewController()?.startAnimating()
             }
         }
-        
+        vc.onEnable2faCall = { [weak self] in
+            if let appDelegate = UIApplication.shared.delegate as? AppDelegate{
+                appDelegate.resolve2faOff()
+            }
+            self?.textContinuation?.resume(throwing: TwoFactorCallError.cancel(localizedDescription: ""))
+
+            DispatchQueue.main.async {
+                let storyboard = UIStoryboard(name: "AuthenticatorFactors", bundle: nil)
+                if let vc = storyboard.instantiateViewController(withIdentifier: "SetPhoneViewController") as? SetPhoneViewController {
+                    vc.phoneCall = true
+                    vc.network = network
+                    vc.isSmsBackup = true
+                    UIApplication.topViewController()?.present(vc, animated: true)
+                }
+            }
+        }
         DispatchQueue.main.async {
             if let appDelegate = UIApplication.shared.delegate as? AppDelegate{
                 appDelegate.resolve2faOn(vc)