+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+- Enable 2FA call method as an SMS backup
+
 ### Fixed
 - Remove lightning account with shortcut